+++ Libc/regex/TRE/lib/tre-match-utils.h
@@ -45,7 +45,7 @@
 	else								      \
 	  {								      \
 	    size_t w;							      \
-	    int max;							      \
+	    size_t max;							      \
 	    if (len >= 0)						      \
 	      max = len - pos;						      \
 	    else							      \
@@ -123,7 +123,7 @@
 	else								      \
 	  {								      \
 	    size_t w;							      \
-	    int max;							      \
+	    size_t max;							      \
 	    if (len >= 0)						      \
 	      max = len - pos;						      \
 	    else							      \
